fix(HTD-77): move helpers to individual folders
